Update patch set 1
Patch Set 1:

Personally I think that's inconsistent with other editors. Either way we should be internally consistent. We will need to have a clickable way to remove nodes to solve the problem in mobile where you don't have access to a delete key, but this might want to be in the context menu.

Patch-set: 1
Label: Verified=0