Use a single unicorn-like <img> instead of &#xFEFF; in inline slugs
Seems to have the same effect for cursor behavior, and does not involve
text that we need the observer to ignore. With this change, typing in
an inline slug works without pawning. The typed text ends up in the
inline slug's <span> before/after the image. It gets SurfaceObserved
correctly despite that, because the <span> isn't treated specially any more.

Changes:
* Replace FEFF with img in inlineSlugTemplate
* Make most slug-specific logic apply only to block slugs
* Account for inline slugs in getNodeAndOffset's unicorn adjustments
* Stop pawning inline slugs; and kill pawning code completely now that it's unused
* Respect render lock in CBN setupSlugs; maybe this should be in onSplice instead?

Tested:
* Typing in empty paragraphs in Chromium
* Typing in between image nodes in Chromium
* Normal unicorning in inline slugs

To test:
* Firefox
* IMEs
* Arrow key navigation

TODO:
* Arrow keying from inline slug onto focusable node into another inline slug
  throws exception
* Aggressively clean up behind inline slugs, like we do for unicorns
* Redo render lock checking in a more carefully considered way

Change-Id: Id360d90d6cf6adad29b1f5f14bc27b7dc8109851