WIP: Implement ve.dm.Transaction.transpose() w/ intentions
Transposing two transactions allows you to reorder them.  It is the basis
of a Operational Transform implementation for Visual Editor.  There is
an extensive test suite ensuring that the result of transposing two
transactions yields consistent results.

This commit adds a functional transpose operation for transactions which
operates on the high level "intention" associated with the transaction,
leaving it to the `ve.dm.Transaction.new*` methods to create operation
sequences which respect document structure.

Based on: https://gerrit.wikimedia.org/r/84442

Change-Id: Iae33bb637a93fd4da09d75d4e01de39e404d1ed9