[VERY WIP] Record intention in ve.dm.Transaction object
For OT purposes we try to reconcile the high level intention of the
transaction, not the raw low-level operations.  As a first step,
record the intention of the transaction in a field when it is created.

Note that the `intention` field is an array consisting of the name of a
constructor method in `ve.dm.Transaction` (starting with "new"), followed
by the non-document arguments to the constructor.  The first argument is
always a `ve.dm.Range` object giving the affected region; some methods
which take numeric offsets have been tweaked to allow collapsed ranges
instead.  If the second argument is a range, it gives the affected
region within the metadata at a given offset (given by a collapsed
range as the first argument).

Based on I71117699b07f78a605d99a568e861d06e8fd3d0e.

Change-Id: Ib678ff105ba532215eff12f340391e32f18b283f