Add an event to updating current values in scalable
The scalable object can have its current values change externally
but this change should propogate to other dependent objects.
Specifically, this is crucial to allow the media dialog to switch
an image without destroying and rebuilding a new scalable, but
rather just updating the current adjusted dimension values
directly, which are then updated in the media size widget.

This allows I4327c86a to work and update dimensions properly.

Change-Id: Idbb8b29db8e47afa9a658a02b2eaf2c3478edb48