Update patch set 4
Patch Set 4:

In MW, the ve.dm.MWInlineImageNode and ve.dm.MWBlockImageNode inhert from the general ve.dm.MWImageNode. 

There's logic about automatic sizing and about scalable and taking its information from the API in the general MWImageNode; we might not use the MW-API for images, but it might be useful to keep some sort of scalable/API logic in a general ve.dm.ImageNode to read the image information from the file or some similar information source.

Shouldn't we keep that structure, and have ImageNode as a parent to both InlineImageNode and BlockImageNode in core as wel?

Patch-set: 4
Reviewer: Mooeypoo <798@e9e9afe9-4712-486d-8885-f54b72dd1951>
Label: Verified=0