Update patch set 1
Patch Set 1: Code-Review-1

There are two issues here:

1. The reason I initialized the dimensions as objects is because otherwise ve.compare() crashed. This might be fixed by setting them to = null, though, which is fine if it works.

2. For the ImageModel and the Scalable to be persistent, the Scalable must have a 'reset' method that emits all the change events. 

The reason is because when the user switches an image in the MediaDialog, we want the Scalable's "base" dimensions (originalDimensions. defaultDimensions and min/max dimensions that all are calculated from the original dimensions, taken from the API) to be purposefully *off* as if the scalable was just created -- *and* to emit the events, so that the MediaSizeWidget and the model can adjust to the fact that these don't exist until the API returns a reply and they are repopulated.

Without this function the MediaSizeWidget kept returning 'invalid' dimensions whenever the dimensions changed between images and was not updating the "full size" button and "default" button in case these dimensions weren't retrieved in time from the API (which caused other annoyance problems).

So "reset" has to exist and must emit all the events. The 'null' issue can be fixed if it doesn't break ve.compare.

You can test this with this I4327c86a9127 commit (which awaits review)

Patch-set: 1
Label: Code-Review=-1