Fix the newline bunny-hop bug harder
In I4375f4c07, I fixed a bug where " <meta> \nfoo" would dirty-diff
to "\n <meta>foo" in wrapper paragraphs. It turns out this was
still broken in the case where the newline was followed by an
inline node rather than text: " <meta>\n<span>" was still being
dirty-diffed to "\n <meta><span>".

This was caused by a special case for content nodes being inserted
into wrappers which clears wrappedWhitespace. In addition to that,
it also needs to output any buffered meta items and their whitespace,
and it needs to clear nextWhitespace so we don't later record
the previous value of wrappedWhitespace as outer whitespace in
the content node's .internal.whitespace array.

Bug: 69861
Change-Id: Ie80a98512b350311832a8674fac892b378738df0