Fix the newline bunny-hop bug harder
In I4375f4c07, I fixed a bug where " <meta>\nfoo" would dirty-diff
to "\n <meta>foo" in wrapper paragraphs. It turns out this was
still broken in the case where the newline was followed by an
inline node rather than text: " <meta>\n<span>" was still being
dirty-diffed to "\n <meta><span>".

This was caused by a special case for when a content node is inserted
into wrappers, in which case wrappedWhitespace is cleared. That's fine
and necessary, but in addition to that, we also need to output any
buffered meta items and their whitespace, and clear nextWhitespace
so we don't later record the previous value of wrappedWhitespace as
outer whitespace in the content node's .internal.whitespace array.

Bug: 69861
Change-Id: Ie80a98512b350311832a8674fac892b378738df0