Update patch set 6
Patch Set 6:

We don't need to map every VE node to an HTML element. Given that images will be convertible to figures it makes sense to give them similar names. Also the Block/Inline prefixes accurately describe the difference between them, whereas the difference between Figure & Image is less clear. In fact what we mean by BlockImage is not an HTML figure, as that is just the frame around the image and caption, where BlockImage contains the src & dimensions.

If we were to follow HTML the FigureNode would just be a frame which contained an ImageNode, but then things get really ugly.

Patch-set: 6