Check IMEs actually fail as expected
Previously, each IME test was skipped if failure was expected. Now, the test
harness checks that a failure actually happens, and that it happens at the
expected point in the input sequence. This way, if a change affects IME as a
a side effect, it will be noticed, whether the side effect is positive,
negative or both.

Change-Id: I67656591a6b1af1f64163338bef5e94a3d3b6b78