Avoid undo when processing a cut event.
The previous cut handler ran an undo using execCommand, to
counteract the browser-issued text deletion on the cut event.

Unfortunately in Chromium 36.x, the cut event handler is processed
(even within the zero-timeout block) before the browser deletes the
selection. This results in the undesirable removal of previously
typed text.

To avoid this, simply prevent the browser from cutting using
preventDefault(), and remove the selection using a transaction.

Also added a comment about why the timeout is still needed - the
reason being that firefox still briefly messes with the selection
even after preventDefault().

Bug: 58724

Change-Id: I68993505b242f1ae5a25a9b3152dbf8926f83b82