Avoid undo when processing a cut event.
The previous cut handler ran an undo using execCommand, to
counteract the browser-issued text deletion on the cut event.

Unfortunately in Chromium 36.x, the cut event handler is processed
(even within the zero-timeout block) before the browser deletes the
selection. This results in the undesirable removal of previously
typed text.

To avoid this, simply prevent the browser from cutting using
preventDefault(), and remove the selection using a transaction.

This removes the need for having a zero-timeout as well.

Bug: 58724

Change-Id: I8f3f608e3f0c1bbfd2cda9543bcf7eeea6fa7c37