Update patch set 4
Patch Set 4: Code-Review-2

This approach has a major flaw. While in Chrome onCopy writes straight to the clipboard with the clipboardData API, we can't do that in FF, so it moves content to copy to the pasteTarget and lets the native event happen. If we cancel the cut event then there's no way for data to get written to the clipboard.

We should be able to use beforeCopy/afterCopy though without preventDefault.

Patch-set: 4
Reviewer: Esanders <658@e9e9afe9-4712-486d-8885-f54b72dd1951>
Label: Code-Review=-2