Update patch set 2
Patch Set 2:

Ok, a bit late but I've found the case this breaks. If the surface is blurred (null selection) and you click somewhere, this functions fires before the the selection is set, so the selection gets reset to [1,1], even if you clicked in a specific place.

Patch-set: 2
Reviewer: Esanders <658@e9e9afe9-4712-486d-8885-f54b72dd1951>
Label: Verified=0