Ignore focusin/focusout events on iframes
In IE11, when the focus moves into an iframe, focusout is
emitted on the element that used to have focus (which is correct)
and focusin is emitted on the iframe (which is incorrect).
Even worse, when the focus moves between two iframes,
focusout is emitted on one and focusin is emitted on the other,
while really from the perspective of the main document
nothing is happening.

This confuses the focus change code in ve.ce.Surface,
so ignore focusin/focusout events emitted on an iframe.

Change-Id: I1236887f927df77e3bc2eaa9ec4ba07a69e3beaa