Split getSelectionRect into bounding and inline
Non-embedded contexts should always use bottom-middle of the bounding
box, unless that doesn't touch highlight, in which case we use the
bottom inline rect and place the context at the right corner of that
(or left in RTL).

Bonus: Rangy doesn't understand RTL so stop using Rangy and calculate
start and end positions ourselves.

Get rid of FocusableNode#getDimensions as the dimensions are
inside getBoundingRect.

Change-Id: I00ab986bacbf4e73714d810bc6ca109a5dd4df02