build: Make headers atop dist files suitable for distribution
We were still prepending the header we use for in-repo files, which
isn't very useful once distributed as a standalone file. Also, change
all in-repo files' headers to be simply the copyright message with a
reference to http://ve.mit-license.org for details.

Also fixed grunt-build which was broken (git-build should not be part
of `grunt build` but `grunt test` – this way, for releases we only need
to run `grunt build`, and it won't include the git commit). This is how
we do this in OOjs.

Change-Id: I99acd6c6a6748b2029bb653836fed00dd8b5d21b