[WIP] build: Make headers atop dist files suitable for distribution
We were still prepending the header we use for in-repo files, which
isn't very useful once distributed as a standalone file.

Added a few examples of what we could do to in-repo file headers:

* Strip them (we only distribute these files as part of the repo,
  which has a LICENSE in its root already, we don't need headers
  on HTML, SVG and PNG files either).

* Copyright with reference to http://ve.mit-license.org

* Just reference to http://ve.mit-license.org (like in the minified
  file header).

Also fixed grunt-build which was broken (git-build should not be
part of grunt-build but grunt-test, that way for releases you
only grunt-build, and it won't include the git commit).
Same as we do in OOjs.

Change-Id: I99acd6c6a6748b2029bb653836fed00dd8b5d21b