Update patch set 2
Patch Set 2:

Yeah, either is fine. Looking at the spec, neither is deprecated or for backwards compatibility. Historically it was on Document with browsers implementing a non-standard alias on the Window. When it made it into the spec the alias was reversed (it relates to the browsing context). While detached documents can have their own selection, they also have their own Window object.

Patch-set: 2