Rangy's dead honey, rangy's dead
The final use of rangy was checking if the native selection is backwards.
What we actually want to do is make up/down arrowing consistent in all
browsers when you have a selection.

The only case that needs fixing is when the browser doesn't support backwards
selections but the dm range is backwards, which we can detect easily.

Change-Id: Ifd5000fc5e667cb9fd6db0e97cf8a03b4da45553