Convert all instances of assert.equal to assert.strictEqual
We don't tolerate == in our code, so we shouldn't use it in our tests.

Also replace notEqual with notStrictEqual, and assert.ok( val ) with
assert.strictEqual( val, true ) where val is generated by a non-native
function and we want to check it actually returns a boolean as well.

Fixed one test in ve.ce.test.js which was comparing x with [x].

Change-Id: I256dce8ef854f2581c7ad1029e55bf068d91508d