Update patch set 9
Patch Set 9:

This is the third major revision of this patch. Some history, copied from IRC:

The first version of the patches (I87645b0ddab29133d81c2a98ce44475dd42bce43 and other patches in the xact-xpose-4 topic) operated directly on the low level operation array.  That was nice because there aren't that many different low level operations, and I can do the transpose blindly without having to clone/look at a complete copy of the document.  But it made me nervous because I also couldn't guarantee anything about what invariants I may or may not be breaking in the document structure.

The second major version (PS6 of this change id) was the version I had at wikimania 2014, and worked on the high-level new* operations.  So now we can rely on the new* constructors to ensure that the actual operations preserve document invariants, although it's harder to ensure that the actual transpose is correct.  I think that's probably the right tradeoff to make though.

But there were too many different new* operations, and they didn't form a complete set -- some operations were missing or non-orthogonal, etc.

So now major version 3 (PS9) introduces intentions as a sort of middle layer which is simpler/more orthogonal than the high-level new* operations, but preserves more of the user's "intention" than looking at the low-level operations.  And that's where we are now.

But if you hate version 3, I could be convinced to go back to version 1, which had a bit of elegance to it. ;)

Currently jenkins doesn't like the patch because I need to go through and fix up the test suite.  (And maybe fix some bugs, we'll see.)

Patch-set: 9
Reviewer: C. Scott Ananian <664@e9e9afe9-4712-486d-8885-f54b72dd1951>
Label: Verified=0