Use start rect instead of bounding rect in ve.ui.Toolbar#onSurfaceViewKeyUp
This is a workaround for the fact that getSelectionBoundingClientRect()
doesn't have handling for Chrome's empty paragraph bug but
getSelectionInlineClientRects() does.

This fixes a bug where if you pressed enter to create a new paragraph,
or backspace to delete the last character in a paragraph, and the
toolbar was floating, the surface would scroll up a bit.

Change-Id: I3d76d2dc17cf0c76aff1fed9d049f1ecfa4e0410