Followup b149390: fix dm.Surface's concept of "selectedNode" to match SurfaceFragment's
In SurfaceFragment, there is only a selectedNode if the selection
precisely contains a node. In Surface, however, the same name
was being used to refer to the node the selection is inside of,
even if the selection doesn't fully contain the node.

This led to problems when ui.DesktopContext assumed that
if an inspector is open and selectedNode is not null, we
must be inspecting a node. This is clearly not true,
because when the link inspector is open, the selection
is entirely within one TextNode. Worse, when the user
clicks out of the inspector but within the same TextNode,
the selectedNode doesn't change, and so the inspector
isn't closed.

Bug: 71373
Change-Id: Ib67fe72a029ef8c706495ab150e8316ae4038a18
(cherry picked from commit 0677c3bae718c191e1fbeb6487930b02f0374e53)