Get rid of trigger-happy slugChange event
slugChange fired on every content change and every selection
change, which is way too often for an event that claims to
be slug-specific. While ce.Surface does need to check for
slug exits on every content and selection change, this should
be done in ce.Surface itself, not by emitting slugChange
any time anything remotely related to slugs could possibly
have potentially happened maybe.

Replace slugChange with slugEnter, which is equivalent
to what slugChange(true) used to be. Split up onSlugChange
into onSlugEnter (the if(newSlug) code path) and updateSlug
(the if(this.slugFragment) code path) and call the latter
from onModelSelect and onSlugEnter.

Also get rid of the range parameter to these events:
onSelectionChange, onContentChange, onSlugEnter and
updateSlug all potentially change the document, so this
range parameter will get out of date. Instead, use the
current selection, which will be updated for document
changes.

The pollOnce call in afterPaste was added to ensure
slugChange got emitted; this can now be removed.

Change-Id: Ie3531f71a1bae39ca57be0e641c71d18bc6aa9f5