Don't rebuild entire paragraph when inserting text in slug location
We already didn't rebuild the paragraph when text was inserted
into an existing text node: instead we resized the text node.
But for cases where the text insertion creates a new text node
(when typing into a slug), we would still rebuild the entire paragraph.

Instead, add an operation to DocumentSynchronizer to insert a new
text node, and use this operation in TransactionProcessor when
a text node insertion is detected.

Now that TextNodes can be spliced into existing ContentBranchNodes
without rebuilds, set ce.TextNode's $element to an empty collection,
so onSplice doesn't try to insert a span that will then be removed
again by renderContents. This allows us to avoid rerendering
when SurfaceObserver generates transactions like these
(onSplice will no longer modify the DOM and renderContents
will notice that the rendering is already up-to-date).

Change-Id: Ia45da472720b54faa70465694c7b7bbbb1569735