Update patch set 2
Patch Set 2:

I think part of the problem is the name of the "setCurrentDimensions" method. It's not *quite* setting current dimensions of the widget, it only changes the dimensions if they are somewhat valid (not numeric)

That method needs to probably be a bit simplified, but I also think we should rename it to what it actually does, maybe more like "setScalableReadyDimensions" or something similar.

Patch-set: 2
Reviewer: Mooeypoo <798@e9e9afe9-4712-486d-8885-f54b72dd1951>
Label: Verified=0