[WIP/POC] Match native content start annotation
Change the DM-level painting behaviour when the cursor is at a content start
followed immediately by annotated text, e.g. <p>|<b>xxx</b>...</p> .

Prior behaviour is to apply no annotations. This leads to a DOM/DM
inconsistency, which persists long enough for the user to observe it as of
the parent changeset (which reduced re-render frequency to help IMEs).

Modified behaviour is to apply any annotations that are present on the right,
except link annotations. This behaviour matches Chromium, but Firefox applies
link annotations too, so in Firefox link annotations are shown in the DOM (in
contradiction to the DM) until the next re-render.

Since browser behaviour differs, it may sometimes be necessary to unicorn the
cursor at content start positions in order to get a consistent experience. We
need to investigate whether this would create cursor directionality issues at
BIDI boundaries when we switch to native cursoring.

Change-Id: I65979fed5a3648887ab70a7c9038f6399dc85c50