POC: setSelection emits if continuation unicorns are needed
For some annotations, such as links, browsers differ from the DM (and from
each other) in their native handling of continuation (i.e. whether text typed
immediately after the annotation gets annotated). Continuation unicorns help
impose consistent rules.

Previously, continuation unicorns were only rendered if the insertion
annotations had changed. This did not work if the old position was inside
a link, and the new position at the end of the link, because a continuation
unicorn is needed in that case. The new behaviour forses unicorns to be
rendered at the end of an annotation that does not use native continuation
rules.

Change-Id: I336e66d2894d131047e818bd224ee232dde4123a