Make setSelection emit activeAnnotationChange if continuation unicorns may be needed
For some annotations, such as links, browsers differ from the DM (and from
each other) in their native handling of continuation (i.e. whether text typed
immediately after the annotation gets annotated). Continuation unicorns help
impose consistent rules.

Previously, continuation unicorns were only rendered if the insertion
annotations had changed. This did not work if the old position was inside
a link, and the new position at the end of the link, because a continuation
unicorn is needed in that case. The new behaviour forces unicorns to be
rendered at the end of an annotation that does not use native continuation
rules.

Also fix a bug whereby complex change annotations came from the character
before the start of the changed text, even if the previous selection was
uncollapsed.

Bug: 72108
Change-Id: I336e66d2894d131047e818bd224ee232dde4123a