Make link inspector not close itself in IE... again
My hacks for making the blur/focus code not notice that
the selection moved seem to be working (but only in VE-MW,
not in VE core), but now the SurfaceObserver's polling
is noticing that the range goes to null instead, and that's
causing inspectors to close themselves at the first poll.

To work around this, stop the observer when an inspector
is opened, and restart it once it's closed. There should
be nicer facilities for this in ce.Surface, but right now
we just want to fix a critical IE bug.

In the future, there should be more advanced logic here
with fake selections, so we can un-isolate inspectors.

Bug: 72392
Change-Id: I0a7c6aed37c1688eb81d4ca0c9ded231f9fd68f1