Unisolate dialogs
We don't really need dialogs to be isolated, because we don't
really need to preserve selections.

Compensate for selection not being preserved by reselecting
the fragment on close in FragmentDialog.

Depends on Id9f6bd2b, Ia498b48d8 and I4a233c0584 in VE-MW,
and on I39df86373 and Id25907c4c41b in oojs-ui.

Change-Id: Id425c56d9d246809962b10ecb88010e191c94410