Update patch set 1
Patch Set 1: Code-Review-1

I think that injecting an argument in the constructor between the first three required arguments and the final list of variadic arguments is the tipping point for this signature - it's time to drop the variadic arguments thing and pass an object instead. The object can have an args property that defaults to an empty array, and addition property, such as this new selection restriction stuff which each default to something sensible internally as well (in this case null).

I realize that the bad constructor signature is my fault, but I think this is the right time to fix it.

Patch-set: 1
Label: Code-Review=-1