TableNode, TableContext: Position row/column overlays better when caption is present
When updating overlays, find out if there's a caption, and if yes
offset the overlays on the captioned side by caption's width/height,
appropriately.

The caption is usually at the top, but this can be overridden using
the 'caption-side' CSS property on the <table/> node. 'top' and
'bottom' from CSS 2 spec are widely supported ('top' being the
default), Firefox also supports 'left' and 'right' (CSS 3).

Minor changes to ve.ui.TableContext.css are because we can no longer
set 'top' or 'left' on .ve-ui-tableContext-rowContext and
.ve-ui-tableContext-colContext (these are used for the offsetting).
colContext was okay, rowContext was done weirdly and now is fixed.

Bug: 72829
Change-Id: I140e63d828ce0432ad804be85f86f53432d84166