Safari copy-paste fix
Always place a nbsp; in the clipboard-key span as otherwise WebKit
browsers won't select it.

Fix clipboard-key detection to be recursive when using pasteTarget.

Fix getClipboardHash to be recursive.

Try to avoid https://code.google.com/p/chromium/issues/detail?id=318925
which causes spaces to be stripped if a newline starts with a node, by
letting the pasteTarget be wider.

Bug: 71718

Change-Id: Iba35ab3a2b382fe8d166bde3f6d6ca16ef40ed12