Update patch set 2
Patch Set 2:

Of the hundreds of function calls triggered by a keystroke I doubt it will be the bottleneck. Sequence#match should be pretty efficient as it works backwards from data[i] and bails as soon there is no match.

As far as IMEs go, you'd have to register a IME candidate which should be pretty easy to avoid.

Patch-set: 2