Wrap inline images in a span ce=false
Chromium native cursoring breaks if an img tag has ce=false directly.
See https://code.google.com/p/chromium/issues/detail?id=440877 .
Wrapping it in a span with ce=false works round this bug.

Cursoring is currently programmatic, so this is a dormant problem; it is fixed
here in preparation for a future change that will bring in native cursoring.

Change-Id: Icc4c4799ffa2733da17e3d79ee03611f3a2781f6