Native backspace and delete
Use native backspace and delete actions.

This is particularly important for delete, which generally removes
an entire grapheme cluster.

Backspace usually removes a single code point (which can, however,
be a surrogate pair of code units).

Still use emulated behaviour if one of the following is true:
* the ctrlKey is down; or
* the selection is expanded; or
* we are directly adjacent to an element in the deletion direction.

N.B. This patch also removes the tests of emulated single
character deletion (which no longer happens).
Character deletion is covered to some extent by the IME tests,
which replay recorded browser behaviour into the CE Surface logic.

What is missing is testing with the actual (as opposed to recorded)
browser behaviour. This would be of some use, to test our
assumptions, but is challenging to set up because it is hard
to trigger an actual native backspace action.

Change-Id: I7f775446675139a0430084b46eda56ddc351721a