Native backspace and delete
Use native backspace and delete actions, provided there is some text in the
direction of the deletion. This is particularly important for delete, which
generally removes an entire grapheme cluster. Backspace usually removes a
single code point (which can be a surrogate pair of code units).

Also make backspace/delete tests asynchronous, to test this.

Change-Id: I7f775446675139a0430084b46eda56ddc351721a