Update patch set 4
Patch Set 4:

Most of those sound like upstream bugs in ooui. Will investigate either way. Bounding rects for text don't include line height, browser bug in all browsers. I've added a bit of padding as per find a replace so it's a bit nicer but not exactly the same. Also technically the selection should render underneath the text as it doesn't change text colour, but that would break on content with a defined background colour.

Patch-set: 4