Update patch set 20
Patch Set 20: Code-Review-1

As mentioned on IRC, there is a problem when moving the cursor in RTL across a sentence that has an inline image in it. Like this:

Something foo<a><img></a>bar here.

When moving in rtl from beginning to end (left in RTL) the cursor skips the image properly, however when returning (right in RTL) from end to start, the cursor gets "stuck" before the image in MW and in the standalone it jumps to the previous paragraph, where it gets stuck (goes back and forth across 1 letter)

Patch-set: 20