Record whitespace on generated empty paragraphs correctly
Whitespace was never recorded on generated: 'empty' paragraphs.

This was almost never noticeable because the linmod output
triggered a separate bug in the data->DOM where whitespace
consistency isn't checked correctly. This meant cases like
<li>\n</li> produced crazy linmod output (with the whitespace
being recorded twice on the li and not at all on the empty paragraph)
but still round-tripped correctly. The only case where
this caused round-trip failures was <body>\n</body>.

Change-Id: Iee23f477714fb56ae6829abb842470b93a4037cc